+++ src/settings/utils.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { FormattedMessage } from 'react-intl';
+
+const NAME_REGEXP = new RegExp(/[^a-zA-Z\d]|^.{8,}$/);
+
+export const validatePrefixSuffixName = ({ name }) => ({
+  name: name?.match(NAME_REGEXP)
+    ? <FormattedMessage id="ui-orders.settings.poNumber.nameValidation" />
+    : undefined,
+});